pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-of-chaos"
version = "0.1.0"
description = "Edge of Chaos — persistent memory, autonomous routines, and intelligent warm-start for Claude Code."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Continuum Contributors"}
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
edge = "edge_of_chaos.cli:main"


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
continuum = ["builtin_skills/**/*.yaml", "builtin_skills/**/*.md"]
